:Debuggging

+++ app/entities/requests/offices/OfficeRequest.scala
@@ -6,14 +6,14 @@ import play.api.data.Forms._
 import play.api.data.format.Formats._
 
 
-case class OfficeRequest(name:String,code:String,parent_office:Integer = null)
+case class OfficeRequest(name:String,code:String,parent_office:Long = null)
 
 object OfficeRequest {
   val form: Form[OfficeRequest] = Form(
     mapping(
       "name" -> of[String]
       , "code" -> of[String]
-      , "parent_office" -> of[Integer]
+      , "parent_office" -> of[Long]
     )(OfficeRequest.apply)(OfficeRequest.unapply)
   )
 }
